extract openExternal
